chore: delete old "ClientRequestInterceptor"
