+++ test/modules/http/compliance/http-req-callback.test.ts
@@ -6,7 +6,7 @@ import { IncomingMessage } from 'node:http'
 import https from 'node:https'
 import { HttpServer, httpsAgent } from '@open-draft/test-server/http'
 import { DeferredPromise } from '@open-draft/deferred-promise'
-import { _ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest/index-new'
+import { ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest'
 import { getRequestOptionsByUrl } from '../../../../src/utils/getRequestOptionsByUrl'
 
 const httpServer = new HttpServer((app) => {
@@ -15,7 +15,7 @@ const httpServer = new HttpServer((app) => {
   })
 })
 
-const interceptor = new _ClientRequestInterceptor()
+const interceptor = new ClientRequestInterceptor()
 interceptor.on('request', ({ request }) => {
   if ([httpServer.https.url('/get')].includes(request.url)) {
     return
