+++ test/modules/http/compliance/http-res-read-multiple-times.test.ts
@@ -8,7 +8,7 @@ import { vi, it, expect, beforeAll, afterEach, afterAll } from 'vitest'
 import http, { IncomingMessage } from 'node:http'
 import { HttpServer } from '@open-draft/test-server/http'
 import { HttpRequestEventMap } from '../../../../src'
-import { _ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest/index-new'
+import { ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest'
 
 const httpServer = new HttpServer((app) => {
   app.get('/user', (req, res) => {
@@ -18,7 +18,7 @@ const httpServer = new HttpServer((app) => {
 
 const resolver = vi.fn<HttpRequestEventMap['request']>()
 
-const interceptor = new _ClientRequestInterceptor()
+const interceptor = new ClientRequestInterceptor()
 interceptor.on('request', resolver)
 
 beforeAll(async () => {
