+++ test/modules/http/compliance/https-constructor.test.ts
@@ -9,14 +9,14 @@ import { URL } from 'node:url'
 import { DeferredPromise } from '@open-draft/deferred-promise'
 import { HttpServer, httpsAgent } from '@open-draft/test-server/http'
 import { getIncomingMessageBody } from '../../../../src/interceptors/ClientRequest/utils/getIncomingMessageBody'
-import { _ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest/index-new'
+import { ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest'
 
 const httpServer = new HttpServer((app) => {
   app.get('/resource', (req, res) => {
     res.status(200).send('hello')
   })
 })
-const interceptor = new _ClientRequestInterceptor()
+const interceptor = new ClientRequestInterceptor()
 
 beforeAll(async () => {
   await httpServer.listen()
