+++ test/modules/http/http-performance.test.ts
@@ -4,7 +4,7 @@
 import { it, expect, beforeAll, afterAll } from 'vitest'
 import { HttpServer } from '@open-draft/test-server/http'
 import { httpGet, PromisifiedResponse, useCors } from '../../helpers'
-import { _ClientRequestInterceptor } from '../../../src/interceptors/ClientRequest/index-new'
+import { ClientRequestInterceptor } from '../../../src/interceptors/ClientRequest'
 
 function arrayWith<V>(length: number, mapFn: (index: number) => V): V[] {
   return new Array(length).fill(null).map((_, index) => mapFn(index))
@@ -31,7 +31,7 @@ const httpServer = new HttpServer((app) => {
   })
 })
 
-const interceptor = new _ClientRequestInterceptor()
+const interceptor = new ClientRequestInterceptor()
 
 interceptor.on('request', ({ request }) => {
   const url = new URL(request.url)
