+++ test/modules/http/intercept/https.get.test.ts
@@ -3,7 +3,7 @@ import https from 'https'
 import { HttpServer } from '@open-draft/test-server/http'
 import { UUID_REGEXP, waitForClientRequest } from '../../../helpers'
 import { HttpRequestEventMap } from '../../../../src/glossary'
-import { _ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest/index-new'
+import { ClientRequestInterceptor } from '../../../../src/interceptors/ClientRequest'
 
 const httpServer = new HttpServer((app) => {
   app.get('/user', (req, res) => {
@@ -12,7 +12,7 @@ const httpServer = new HttpServer((app) => {
 })
 
 const resolver = vi.fn<HttpRequestEventMap['request']>()
-const interceptor = new _ClientRequestInterceptor()
+const interceptor = new ClientRequestInterceptor()
 interceptor.on('request', resolver)
 
 beforeAll(async () => {
